requirements.txt
# ─── API Framework ───────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
pgvector>=0.3.2
alembic>=1.13.1

# ─── AI / LLM ────────────────────────────────────────────────────────────────
openai>=1.30.0
tiktoken>=0.7.0

# ─── Config / Env ────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# ─── Serialization ───────────────────────────────────────────────────────────
pyyaml>=6.0.1

# ─── Utilities ───────────────────────────────────────────────────────────────
python-multipart>=0.0.9
